Split out consumer and producer.
